convert funnel property correlation table
