+++ frontend/src/scenes/funnels/funnelPropertyCorrelationLogic.test.ts
@@ -1,21 +1,19 @@
-import { funnelLogic } from './funnelLogic'
-import { MOCK_DEFAULT_TEAM, MOCK_TEAM_ID } from 'lib/api.mock'
-import { expectLogic } from 'kea-test-utils'
-import { initKeaTests } from '~/test/init'
-import { preflightLogic } from 'scenes/PreflightCheck/preflightLogic'
-import { eventUsageLogic } from 'lib/utils/eventUsageLogic'
-import { insightLogic } from 'scenes/insights/insightLogic'
-import { AvailableFeature, CorrelationConfigType, InsightLogicProps, InsightShortId, InsightType } from '~/types'
+import { expectLogic, partial } from 'kea-test-utils'
+import { MOCK_DEFAULT_TEAM } from 'lib/api.mock'
 import { teamLogic } from 'scenes/teamLogic'
 import { userLogic } from 'scenes/userLogic'
-import { router } from 'kea-router'
-import { urls } from 'scenes/urls'
-import { useMocks } from '~/mocks/jest'
 import { useAvailableFeatures } from '~/mocks/features'
-import api from 'lib/api'
-import { openPersonsModal } from 'scenes/trends/persons-modal/PersonsModal'
-
-jest.mock('scenes/trends/persons-modal/PersonsModal')
+import { useMocks } from '~/mocks/jest'
+import { initKeaTests } from '~/test/init'
+import {
+    AvailableFeature,
+    CorrelationConfigType,
+    FunnelCorrelationResultsType,
+    InsightLogicProps,
+    InsightShortId,
+    InsightType,
+} from '~/types'
+import { DEFAULT_EXCLUDED_PERSON_PROPERTIES, funnelPropertyCorrelationLogic } from './funnelPropertyCorrelationLogic'
 
 const Insight12 = '12' as InsightShortId
 const Insight123 = '123' as InsightShortId
@@ -122,9 +120,10 @@ const funnelResults = [
     },
 ]
 
-describe('funnelLogic', () => {
-    let logic: ReturnType<typeof funnelLogic.build>
-    const correlationConfig: CorrelationConfigType = {}
+describe('funnelPropertyCorrelationLogic', () => {
+    const props = { dashboardItemId: Insight123, syncWithUrl: true }
+    let logic: ReturnType<typeof funnelPropertyCorrelationLogic.build>
+    let correlationConfig: CorrelationConfigType = {}
 
     beforeEach(() => {
         useAvailableFeatures([AvailableFeature.CORRELATION_ANALYSIS, AvailableFeature.GROUP_ANALYTICS])
@@ -317,7 +316,7 @@ describe('funnelLogic', () => {
             },
         })
         initKeaTests(false)
-        window.POSTHOG_APP_CONTEXT = undefined // to force API request to /api/project/@current
+        // window.POSTHOG_APP_CONTEXT = undefined // to force API request to /api/project/@current
     })
 
     const defaultProps: InsightLogicProps = {
@@ -335,294 +334,113 @@ describe('funnelLogic', () => {
         },
     }
 
-    async function initFunnelLogic(props: InsightLogicProps = defaultProps): Promise<void> {
+    async function initPropertyFunnelCorrelationLogic(props: InsightLogicProps = defaultProps): Promise<void> {
         teamLogic.mount()
         await expectLogic(teamLogic).toFinishAllListeners()
         userLogic.mount()
         await expectLogic(userLogic).toFinishAllListeners()
-        logic = funnelLogic(props)
+        logic = funnelPropertyCorrelationLogic(props)
         logic.mount()
         await expectLogic(logic).toFinishAllListeners()
     }
 
-    describe('core assumptions', () => {
-        beforeEach(async () => {
-            await initFunnelLogic()
+    it('Selecting all properties returns expected result', async () => {
+        await initPropertyFunnelCorrelationLogic(props)
+        await expectLogic(logic, () => logic.actions.setPropertyNames(logic.values.allProperties))
+            .toFinishListeners()
+            .toMatchValues({
+                propertyCorrelations: {
+                    events: [
+                        {
+                            event: { event: 'some property' },
+                            success_count: 1,
+                            failure_count: 1,
+                            odds_ratio: 1,
+                            correlation_type: 'success',
+                            result_type: FunnelCorrelationResultsType.Properties,
+                        },
+                        {
+                            event: { event: 'another property' },
+                            success_count: 1,
+                            failure_count: 1,
+                            odds_ratio: 1,
+                            correlation_type: 'failure',
+                            result_type: FunnelCorrelationResultsType.Properties,
+                        },
+                    ],
+                },
+            })
+    })
+
+    it('Deselecting all returns empty result', async () => {
+        await initPropertyFunnelCorrelationLogic(props)
+        await expectLogic(logic, () => logic.actions.setPropertyNames([]))
+            .toDispatchActions(logic, ['loadPropertyCorrelationsSuccess'])
+            .toMatchValues({
+                propertyCorrelations: {
+                    events: [],
+                },
+            })
+    })
+
+    it('isPropertyExcludedFromProject returns true initially, then false when excluded, and is persisted to team config', async () => {
+        await initPropertyFunnelCorrelationLogic(props)
+
+        expect(logic.values.isPropertyExcludedFromProject('some property')).toBe(false)
+
+        await expectLogic(logic, () => logic.actions.excludePropertyFromProject('some property')).toFinishAllListeners()
+
+        expect(logic.values.isPropertyExcludedFromProject('some property')).toBe(true)
+
+        await expectLogic(teamLogic).toMatchValues({
+            currentTeam: partial({
+                correlation_config: {
+                    excluded_person_property_names: DEFAULT_EXCLUDED_PERSON_PROPERTIES.concat(['some property']),
+                },
+            }),
         })
 
-        it('mounts all sorts of logics', async () => {
-            await expectLogic(logic).toMount([
-                eventUsageLogic,
-                insightLogic({ dashboardItemId: undefined }),
-                preflightLogic,
-            ])
-            await expectLogic(preflightLogic).toDispatchActions(['loadPreflightSuccess'])
-        })
+        // Also make sure that excluding the property again doesn't double
+        // up on the config list
+        await expectLogic(logic, () => logic.actions.excludePropertyFromProject('some property')).toFinishAllListeners()
 
-        it('sets filters after load if valid', async () => {
-            await expectLogic(logic)
-                .toDispatchActions(['loadResults'])
-                .toMatchValues({
-                    insight: expect.objectContaining({
-                        short_id: undefined,
-                        filters: {
-                            insight: InsightType.FUNNELS,
-                            actions: [
-                                { id: '$pageview', order: 0 },
-                                { id: '$pageview', order: 1 },
-                            ],
-                        },
-                        result: null,
-                    }),
-                    filters: {
-                        insight: InsightType.FUNNELS,
-                        actions: [
-                            { id: '$pageview', order: 0 },
-                            { id: '$pageview', order: 1 },
-                        ],
-                    },
-                    isFunnelWithEnoughSteps: true,
-                })
-                .toDispatchActions(['loadResultsSuccess'])
-                .toMatchValues({
-                    insight: expect.objectContaining({
-                        filters: {
-                            insight: InsightType.FUNNELS,
-                            actions: [
-                                { id: '$pageview', order: 0 },
-                                { id: '$pageview', order: 1 },
-                            ],
-                        },
-                        result: expect.arrayContaining([expect.objectContaining({ count: 19 })]),
-                    }),
-                    filters: {
-                        insight: InsightType.FUNNELS,
-                        actions: [
-                            { id: '$pageview', order: 0 },
-                            { id: '$pageview', order: 1 },
-                        ],
-                    },
-                    isFunnelWithEnoughSteps: true,
-                })
+        await expectLogic(teamLogic).toMatchValues({
+            currentTeam: partial({
+                correlation_config: {
+                    excluded_person_property_names: DEFAULT_EXCLUDED_PERSON_PROPERTIES.concat(['some property']),
+                },
+            }),
         })
     })
 
-    describe('isFunnelWithEnoughSteps', () => {
-        beforeEach(async () => {
-            await initFunnelLogic()
+    it('loads property exclude list from Project settings', async () => {
+        correlationConfig = { excluded_person_property_names: ['some property'] }
+        await initPropertyFunnelCorrelationLogic(props)
+
+        await expectLogic(teamLogic).toMatchValues({
+            currentTeam: partial({
+                correlation_config: { excluded_person_property_names: ['some property'] },
+            }),
         })
-
-        it('sets it properly', () => {
-            expectLogic(logic, () => {
-                logic.actions.setFilters({ actions: [] })
-            }).toMatchValues({ isFunnelWithEnoughSteps: false })
-
-            expectLogic(logic, () => {
-                logic.actions.setFilters({})
-            }).toMatchValues({ isFunnelWithEnoughSteps: false })
-
-            expectLogic(logic, () => {
-                logic.actions.setFilters({ actions: [{}, {}] })
-            }).toMatchValues({ isFunnelWithEnoughSteps: true })
-
-            expectLogic(logic, () => {
-                logic.actions.setFilters({ events: [{}, {}] })
-            }).toMatchValues({ isFunnelWithEnoughSteps: true })
-
-            expectLogic(logic, () => {
-                logic.actions.setFilters({ events: [{}], actions: [{ from: 'previous isFunnelWithEnoughSteps test' }] })
-            }).toMatchValues({ isFunnelWithEnoughSteps: true })
-        })
-    })
-
-    it("load results, don't send breakdown if old visualisation is shown", async () => {
-        jest.spyOn(api, 'createResponse')
-        await initFunnelLogic()
-
-        // wait for clickhouse features to be enabled, otherwise this won't call "loadResults"
-        await expectLogic(preflightLogic).toDispatchActions(['loadPreflightSuccess'])
 
         await expectLogic(logic, () => {
-            logic.actions.setFilters({
-                actions: [],
-                events: [
-                    { id: '$pageview', order: 0 },
-                    { id: '$pageview', order: 1 },
-                    { id: '$pageview', order: 2 },
-                ],
-                breakdown: '$active_feature_flags',
-            })
+            logic.actions.setPropertyNames(logic.values.allProperties)
+            // logic.actions.loadResultsSuccess({ filters: { insight: InsightType.FUNNELS } })
         })
-            .toDispatchActions(['setFilters', 'loadResults', 'loadResultsSuccess'])
+            .toFinishAllListeners()
             .toMatchValues({
-                apiParams: expect.objectContaining({
-                    actions: [],
+                propertyCorrelations: {
                     events: [
-                        { id: '$pageview', order: 0 },
-                        { id: '$pageview', order: 1 },
-                        { id: '$pageview', order: 2 },
+                        {
+                            event: { event: 'another property' },
+                            success_count: 1,
+                            failure_count: 1,
+                            odds_ratio: 1,
+                            correlation_type: 'failure',
+                            result_type: FunnelCorrelationResultsType.Properties,
+                        },
                     ],
-                    breakdown: undefined,
-                    breakdown_type: undefined,
-                }),
-            })
-
-        expect(api.createResponse).toHaveBeenNthCalledWith(
-            2,
-            `api/projects/${MOCK_TEAM_ID}/insights/funnel/`,
-            expect.objectContaining({
-                actions: [],
-                events: [
-                    { id: '$pageview', order: 0 },
-                    { id: '$pageview', order: 1 },
-                    { id: '$pageview', order: 2 },
-                ],
-                breakdown: undefined,
-                breakdown_type: undefined,
-                insight: 'FUNNELS',
-            }),
-            expect.anything()
-        )
-    })
-
-    describe('syncs with insightLogic', () => {
-        const props = { dashboardItemId: Insight123 }
-        beforeEach(async () => {
-            await initFunnelLogic(props)
-        })
-
-        it('setFilters calls insightLogic.setFilters', async () => {
-            await expectLogic(logic, () => {
-                logic.actions.setFilters({ insight: InsightType.FUNNELS, events: [{ id: 42 }] })
-            })
-                .toDispatchActions([
-                    (action) =>
-                        action.type === insightLogic(props).actionTypes.setFilters &&
-                        action.payload.filters?.events?.[0]?.id === 42,
-                ])
-                .toMatchValues(logic, {
-                    filters: expect.objectContaining({
-                        events: [{ id: 42 }],
-                    }),
-                })
-                .toMatchValues(insightLogic(props), {
-                    filters: expect.objectContaining({
-                        events: [{ id: 42 }],
-                    }),
-                })
-        })
-
-        it('insightLogic.setFilters updates filters', async () => {
-            await expectLogic(logic, () => {
-                insightLogic(props).actions.setFilters({ insight: InsightType.FUNNELS, events: [{ id: 42 }] })
-            })
-                .toMatchValues(logic, {
-                    filters: expect.objectContaining({
-                        events: [{ id: 42 }],
-                    }),
-                })
-                .toMatchValues(insightLogic(props), {
-                    filters: expect.objectContaining({
-                        events: [{ id: 42 }],
-                    }),
-                })
-        })
-    })
-
-    describe('it opens the PersonsModal', () => {
-        const props = { dashboardItemId: Insight123 }
-        beforeEach(async () => {
-            await initFunnelLogic(props)
-        })
-
-        test('openPersonsModalForStep calls openPersonsModal', async () => {
-            await expectLogic().toDispatchActions(preflightLogic, ['loadPreflightSuccess'])
-            await expectLogic(() => {
-                router.actions.push(urls.insightEdit(Insight123))
-            })
-
-            logic.actions.openPersonsModalForStep({
-                step: {
-                    action_id: '$pageview',
-                    average_conversion_time: 0,
-                    median_conversion_time: 0,
-                    count: 1,
-                    name: '$pageview',
-                    order: 0,
-                    type: 'events',
-                    // Breakdown must be ignored in openPersonsModalForStep
-                    converted_people_url: '/some/people/url?funnel_step=2&funnel_step_breakdown=USA',
-                    dropped_people_url: '/some/people/url?funnel_step=-2&funnel_step_breakdown=USA',
                 },
-                converted: true,
             })
-
-            expect(openPersonsModal).toHaveBeenCalledWith({
-                title: expect.any(Object),
-                url: '/some/people/url?funnel_step=2', // Positive funnel_step and no funnel_step_breakdown
-            })
-        })
-
-        test('openPersonsModalForSeries calls openPersonsModal', async () => {
-            await expectLogic().toDispatchActions(preflightLogic, ['loadPreflightSuccess'])
-            await expectLogic(() => {
-                router.actions.push(urls.insightEdit(Insight123))
-            })
-
-            logic.actions.openPersonsModalForSeries({
-                series: {
-                    action_id: '$pageview',
-                    average_conversion_time: 0,
-                    median_conversion_time: 0,
-                    count: 1,
-                    name: '$pageview',
-                    order: 0,
-                    type: 'events',
-                    // Breakdown must be ignored in openPersonsModalForStep
-                    converted_people_url: '/some/people/url?funnel_step=2&funnel_step_breakdown=Latvia',
-                    dropped_people_url: '/some/people/url?funnel_step=-2&funnel_step_breakdown=Latvia',
-                    droppedOffFromPrevious: 0,
-                    conversionRates: {
-                        fromPrevious: 1,
-                        total: 1,
-                        fromBasisStep: 1,
-                    },
-                },
-                step: {
-                    action_id: '$pageview',
-                    average_conversion_time: 0,
-                    median_conversion_time: 0,
-                    count: 1,
-                    name: '$pageview',
-                    order: 0,
-                    type: 'events',
-                    // Breakdown must be ignored in openPersonsModalForStep
-                    converted_people_url: '/some/people/url?funnel_step=2&funnel_step_breakdown=USA',
-                    dropped_people_url: '/some/people/url?funnel_step=-2&funnel_step_breakdown=USA',
-                },
-                converted: true,
-            })
-
-            expect(openPersonsModal).toHaveBeenCalledWith({
-                title: expect.any(Object),
-                url: '/some/people/url?funnel_step=2&funnel_step_breakdown=Latvia', // Series funnel_step_breakdown included
-            })
-        })
-    })
-
-    describe('funnel simple vs. advanced mode', () => {
-        beforeEach(async () => {
-            await initFunnelLogic()
-        })
-
-        it("toggleAdvancedMode() doesn't trigger a load result", async () => {
-            await expectLogic(logic, () => {
-                logic.actions.toggleAdvancedMode()
-            })
-                .toDispatchActions(['toggleAdvancedMode', 'setFilters'])
-                .toNotHaveDispatchedActions([
-                    insightLogic({ dashboardItemId: Insight123 }).actionCreators.loadResults(),
-                ])
-        })
     })
 })
